Separate GCSEvent into top-level class for reuseability.
